Changes #6
1) Implemented group ordering for user in Requests page
2) Implemented requests displaying for the specefic group
3) Request page can be opened from requests page list table
4) Changed some coloumns in DB user table
5) Now user default request group saved in session and DB
6) Added new String value to Beans
7) Fixed bug to show necessary user group

+++ src/org/RequestDesk/controller/LoginController.java
@@ -70,7 +70,7 @@ public class LoginController extends HttpServlet
 		        session.setAttribute("fullname", user.GetFullName());
 		        session.setAttribute("lastlogindate", user.GetLastLoginDate());
 		        session.setAttribute("language", user.GetLanguage());
-		        session.setAttribute("requestsgroup", user.GetRequestsGroup());
+		        session.setAttribute("requestsgroup", (Integer) user.GetRequestsGroup());
 		        
 		        Cookie cookieLogin = new Cookie("login", user.GetLogin());
 		        Cookie cookieSession = new Cookie("session", user.GetSession());
